Update main.cpp
changed pitch bend behavior, now with the intent of using it as a separate CV output. 

also added re-trigger functionality to the gate output. handleNoteOn() now turns off the gate for 32 microseconds, and sendGate turns it back on. this behavior can be disabled by setting the RETRIGGER constant to false. (there may be a better method than using delayMicroseconds, but i haven't noticed any bad behavior or missed MIDI.)